Add files via upload
Customer LIfe Time Value:
Our client is a major non-life insurance company that wants to evaluate customer life time value based on each customer's demographics and policy information including claim details. The CLV is a profitablity metric the organization has set in terms of a value placed by the company on each customer and can be conceived in two dimensions: the customer`s present Value and potential future Value.

It is important for the company to take an optimal decision and implement appropriate action plans. They will be able to do this by accurately predicting the CLV of its customers.

An analytical and modelling framework to predict the life time value of each customer is designed.Various statistical and machine learning models were applied to predict the CLV. This is based on the quantitative and qualitative features provided in the dataset